requirements.txt
# Core data processing
pandas
numpy

# Machine Learning / Deep Learning
torch
transformers
sentence-transformers
scikit-learn

# Similarity search
faiss-cpu

# App / UI
gradio
streamlit

# LLM orchestration
langchain
langchain-google-genai

# Image handling
Pillow
opencv-python

# Utils
tqdm
python-dotenv
